+++ CHANGELOG.rst
@@ -2,9 +2,16 @@
 Version History
 ===============
 
+v6.11.3
+-------
+
+* Remove loop in group subscription state machine. `<https://github.com/lsst-ts/LOVE-frontend/pull/787>`_
+* Temporary disable control location feature. `<https://github.com/lsst-ts/LOVE-frontend/pull/780>`_
+
 v6.11.2
 -------
 
+* Update package.json version to v6.11.2. `<https://github.com/lsst-ts/LOVE-frontend/pull/781>`_
 * Refactor and fix ContextMenu logic in ScriptQueue component. `<https://github.com/lsst-ts/LOVE-frontend/pull/777>`_
 * Add ScriptQueue component property to control wether to display or not the observatory status.  `<https://github.com/lsst-ts/LOVE-frontend/pull/776>`_
 * Add toggle button to the ScriptQueue ConfigPanel component to allow bypassing schema verifications. `<https://github.com/lsst-ts/LOVE-frontend/pull/778>`_